#25536 moved flotiq setup command logic to module
