+++ package.json
@@ -2,12 +2,12 @@
   "name": "flotiq-setup",
   "version": "1.0.1",
   "description": "Flotiq-setup makes it easier to connect your projects to Flotiq",
-  "main": "index.mjs",
+  "main": "index.js",
   "bin": {
-    "flotiq-setup": "./index.mjs"
+    "flotiq-setup": "./index.js"
   },
   "scripts": {
-    "start": "node index.mjs"
+    "start": "node index.js"
   },
   "keywords": [
     "flotiq",
@@ -18,10 +18,12 @@
   "dependencies": {
     "@types/yargs": "^17.0.32",
     "axios": "^1.6.8",
-    "chalk": "^5.3.0",
+    "chalk": "^4.1.1",
     "dotenv": "^16.4.5",
+    "esm": "^3.2.25",
     "express": "^4.19.2",
     "open": "^10.1.0",
+    "openurl": "^1.1.1",
     "yargs": "^17.7.2"
   }
 }
